test: move ping timeout test to standalone PingTimeoutTest class
ClientTest.setUp() requires a live ClickHouse connection, so the timeout
test (which only needs a local socket server) is extracted into its own
class to run without any external dependencies.
